refactor(qauld-ctl): extract qauld-rpc shared crate
Move the RPC plumbing into a new workspace member at rust/qauld-rpc/
so a future TUI binary can depend on it without forking qauld-ctl.

Migrated:
  - RpcCommand trait (the protocol-level contract)
  - RpcTransport trait + SocketTransport + EmbeddedTransport
  - QaulRpc envelope encoding helpers
  - id_string_to_bin / uuid_string_to_bin

qauld-ctl now re-exports RpcCommand and the helpers from qauld-rpc
to keep existing intra-crate import paths working. Per-subcommand
RpcCommand impls (clap-derived) stay in qauld-ctl.

The 'embedded' feature on qauld-ctl now forwards to qauld-rpc's
matching feature. Default and --features embedded builds verified;
end-to-end regression smoke against a fresh daemon still works.
